backend/requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
numpy
torch
openai-whisper
librosa
sounddevice